refactor(phase-2k): extract OpenProjectProjectService (project reads)
Continues the openproject_client.py god-class decomposition (ADR-002).
The four project-read helpers — full-list, single by identifier,
enhanced-with-stats, batched lookup — move into a new focused
``OpenProjectProjectService`` exposed as ``self.projects``.
``OpenProjectClient`` keeps a thin delegator for each method so
existing callers work unchanged.

Methods moved
-------------
- ``get_projects`` (file-based read with rails-runner fallback for
  console instability)
- ``get_project_by_identifier`` (lazy-imports ``escape_ruby_single_quoted``
  to avoid the service ↔ client cycle)
- ``get_project_enhanced`` (project + work-package count via JSON query)
- ``batch_get_projects_by_identifiers`` — the ``@batch_idempotent``
  decorator travels with the moved method, and the signature gains
  ``headers: dict[str, str] | None = None`` so callers who actually
  want cached results can supply ``X-Idempotency-Key`` (consistent
  with the Phase 2j ``batch_get_users_by_emails`` fix)

Project *write* helpers (``upsert_project_attribute``, the modules
family, ``ensure_project_version``) stay on the client for now and
earn their own follow-up extractions (Phase 2k.2 / 2k.3) — they have
heavier coupling to the work-package CF subsystem and the migration
runner.

Numbers
-------
- ``openproject_client.py``: 5,103 → 4,829 LOC (-274)
- ``openproject_project_service.py``: 0 → 392 LOC (new)
- Cumulative across phases 2a–2k: ``openproject_client.py``
  7,342 → 4,829 LOC (-2,513, -34.2%)

Verification
------------
- pytest tests/unit: 953 passed
- mypy src/: clean (103 files)
- ruff check / ruff format: clean
